Reorg app a bit to start using more security features
