+++ app/index.html
@@ -22,8 +22,8 @@
         <script src="snap/cloud.js"></script>
         <script src="snap/sha512.js"></script>
 
-        <script src="snap-app/morphic.js"></script>
-        <script src="snap-app/gui.js"></script>
+        <script src="snap-proxies/morphic.js"></script>
+        <script src="snap-proxies/gui.js"></script>
 
         <script type="text/javascript">
             var world;
